UI Update + Button Static Mesh
Created the UI For the value displays and the reticle. Note that these are unbound as of know, but will be bound when their functionality in the Game Instance is created.

Also added the static mesh for the button so that it can be used for the creation of the button and its many subclasses