Began work on User Interface and VR Compatibility
In order to fix the UI, we'll need to render the components in 3D space, which means we will have to update the requirements / design document. I have done the Value Displays as a prototype, but we will need to fix the rest of the UI elements. As a side effect, we now have more control over the UI elements in VR, meaning we can rotate and manipulate them however we like.

Also made it so that the player character rises 150 units if in VR