VR UI Always visible
This fix allows 3D UI To be seen everywhere, regardless of obstruction. To achieve this, the material used by the depth pass was set to disable depth test, making it render regardless of visibility. In order to make use of this, all UI elements are now rendered in translucent mode, which makes them blurry in movement, but I think this can be fixed by moving the UI further from the player and scaling it to be larger. We'll need to experiment.