Added RaiseFlag() Function
Added the raiseFlag() function to the button. This is where the buttons will raise their appropriate flags / carry out actions that they need to take. This was added as a lot of the button actions will result in buttons being deleted, thus the animation would never be seen by the user. It is important to have the animation play to give feedback to the player, which makes the simulation feel more fluid.