Filled HandleFlags function in Narator.cpp
Added include files for all buttons
Created a for loop to iterate through list of buttons and to check if the flag was raised
Attempted to implement casting to specific buttons but will need some work. Buttons are also yet to have their respective functions so there is no real way to test to see if it works.