Destination Button Testing
-Made the mapName field on the destination button editAnywhere so it could be modified in editor, and thus tested independently of the narrator
-Added a destination button to the Minimal_Default map. When clicked, it takes the player to the Plane map.

NOTE: The mapName field has to be identical to the name of the level we travel to, and we don't have any way to allias map names with human readable names. Thus for now, we need to keep map names as human readable (And without spaces), or we need to make a lookup function in the game instance that associates map names with human readable names.